refactor: movieInfoWithTitle 타입 성성
