+++ src/hooks/useGetMovieImgAndTitle.tsx
@@ -1,11 +1,10 @@
 import { getMovieTopRated } from '@apis/getMovieNowPlaying';
+import { movieInfowithTitle } from 'types/movie';
 
 export default async function useGetMovieImgAndTitleAndId() {
-  const getMovieTopRatedImgAndTitle: {
-    poster_path: string;
-    title: string;
-    id: string;
-  }[] = (await getMovieTopRated()).map((movie: any) => ({
+  const getMovieTopRatedImgAndTitle: movieInfowithTitle[] = (
+    await getMovieTopRated()
+  ).map((movie: any) => ({
     poster_path: movie.poster_path,
     title: movie.title,
     id: movie.id,
