+++ src/types/movie.ts
@@ -1,0 +1,10 @@
+export interface movieInfo {
+  posterPath: string;
+  movieId: number;
+}
+
+export interface movieInfowithTitle {
+  poster_path: string;
+  title: string;
+  id: string;
+}
