--- src/types/movie.tsx
@@ -1,4 +1,0 @@
-export interface movieInfo {
-  posterPath: string;
-  movieId: number;
-}
